packages/ni.datamonikers.v1.client/pyproject.toml
[project]
name = "ni.datamonikers.v1.client"
version = "0.1.0-dev0"
license = "MIT"
description = "gRPC Client for NI Data Moniker Service"
authors = [{name = "NI", email = "opensource@ni.com"}]
maintainers = [
  {name = "Joe Friedrichsen", email = "joe.friedrichsen@emerson.com"},
  {name = "Brad Keryan", email = "brad.keryan@emerson.com"}
]
readme = "README.md"
keywords = ["data-moniker", "client"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Intended Audience :: Manufacturing",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Operating System :: Microsoft :: Windows",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: 3.13",
  "Programming Language :: Python :: Implementation :: CPython",
]
dynamic = ["dependencies"]
requires-python = '>=3.9,<4.0'

[project.urls]
repository = "https://github.com/ni/ni-apis-python"

[tool.poetry]
packages = [{include = "ni", from = "src"}]
requires-poetry = '>=2.1,<3.0'

[tool.poetry.dependencies]
ni-grpc-extensions = { version = ">=0.1.0.dev0", allow-prereleases = true }
ni-datamonikers-v1-proto = { version = ">=0.1.0.dev0", allow-prereleases = true }

[tool.poetry.group.dev.dependencies]
types-grpcio = ">=1.0"
types-protobuf = ">=4.21"

[tool.poetry.group.lint.dependencies]
bandit = { version = ">=1.7", extras = ["toml"] }
ni-python-styleguide = ">=0.4.1"
mypy = ">=1.0"
pyright = { version = ">=1.1.400", extras = ["nodejs"] }

[tool.poetry.group.test.dependencies]
pytest = ">=7.2"
pytest-cov = ">=4.0"
pytest-doctestplus = ">=1.4"
pytest-mock = ">=3.0"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
# The latest Sphinx requires a recent Python version.
Sphinx = [
  { version = ">=7.4", python = ">=3.9,<3.10" },
  { version = ">=8.1", python = ">=3.10,<3.11" },
  { version = ">=8.2", python = "^3.11" },
]
sphinx-rtd-theme = ">=1.0.0"
sphinx-autoapi = ">=1.8.4"
m2r2 = ">=0.3.2"
toml = ">=0.10.2"

[build-system]
requires = ["poetry-core>=2.1.0,<3.0"]
build-backend = "poetry.core.masonry.api"


[tool.bandit]
skips = [
  "B101", # assert_used
]

[tool.ni-python-styleguide]
extend_exclude = "docs"
application-import-names = "ni.datamonikers.v1.client"

[tool.black]
extend-exclude = 'docs/'
line-length = 100

[tool.mypy]
mypy_path = "src"
files = "."
namespace_packages = true
strict = true
explicit_package_bases = true
exclude = ["docs"]

[tool.pyright]
include = ["src/", "tests/"]

[tool.pytest.ini_options]
addopts = "--doctest-modules --doctest-plus --strict-markers"
testpaths = ["tests"]
